+++ CHANGELOG.md
@@ -7,15 +7,19 @@
 ### FEATURES
 
 - Add support for streamlining releases ([#507])
+- [relayer-cli]
+  - Implement command to query the channels associated with a connection ([#505])
 
 ### IMPROVEMENTS
 
 - Update to `tendermint-rs` v0.17.1 ([#517])
 
 
+[#505]: https://github.com/informalsystems/ibc-rs/issues/505
 [#507]: https://github.com/informalsystems/ibc-rs/issues/507
 [#511]: https://github.com/informalsystems/ibc-rs/pull/511
 [#517]: https://github.com/informalsystems/ibc-rs/issues/517
+
 
 ## v0.0.6
 *December 23, 2020*